+++ relay-server/src/services/processor.rs
@@ -298,8 +298,12 @@ impl ProcessingGroup {
         }
 
         // Extract spans.
-        let span_items = envelope
-            .take_items_by(|item| matches!(item.ty(), &ItemType::Span | &ItemType::OtelSpan));
+        let span_items = envelope.take_items_by(|item| {
+            matches!(
+                item.ty(),
+                &ItemType::Span | &ItemType::OtelSpan | &ItemType::OtelTracesData
+            )
+        });
         if !span_items.is_empty() {
             grouped_envelopes.push((
                 ProcessingGroup::Span,
@@ -1856,6 +1860,7 @@ impl EnvelopeProcessorService {
         state: &mut ProcessEnvelopeState<SpanGroup>,
     ) -> Result<(), ProcessingError> {
         span::filter(state);
+        span::convert_otel_traces_data(state);
 
         if_processing!(self.inner.config, {
             let global_config = self.inner.global_config.current();